discogs_credits: extract UI bar + entry slim-down + missing-import fixes
Issue #32 — last big move:

  src/ui-bar.js
    insertDiscogsBar (the import bar with embedded CSS) + startImportRels
    (the post-click orchestrator that runs preflight → review table →
    dispatch) + the localStorage-cleanup IIFE. Module-private `_logs` /
    `_summary` state replaces the entry's `let logs, summary, discogsUrl`.

  src/discogs_credits.user.js
    Rewritten as a 70-line bootstrap: just the cross-tab BroadcastChannel
    listener + the `$(document).ready` that calls `hasDiscogsLinkDefined`
    then `insertDiscogsBar`. Side-effect imports `./storage.js` (opens
    IDB) and `./ui-bar.js` (runs the cleanup IIFE).

Bug fixes that came out of the move:

  - api-discogs.js gains `clearReleaseDataCache(url)` — the entry had a
    stale `_releaseDataCache.delete(discogsUrl)` reference left over
    from chunk 2c that would have thrown if the user clicked Refresh.
  - ui-bar.js was missing imports for `ENTITY_TYPE_MAP`,
    `getAllArtistTracks`, `getArtistRoles`; review-table.js was missing
    `fetchWithRetry`. Without these the preflight hung on first run.
    Surfaced by smoke timeout + browser console "ENTITY_TYPE_MAP is not
    defined".

Smoke (Spiritual Jazz: 95 staged rels) green.
